requirements.txt
async-generator==1.10
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
Flask==2.0.2
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-language==2.3.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
outcome==1.1.0
pandas==1.3.4
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.8
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
webdriver-manager==3.5.2
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
